+++ examples/data-sources/iosxr_router_ospf_vrf_redistribute_bgp/data-source.tf
@@ -1,5 +1,0 @@
-data "iosxr_router_ospf_vrf_redistribute_bgp" "example" {
-  process_name = "OSPF1"
-  vrf_name     = "VRF1"
-  as_number    = "65001"
-}