+++ examples/resources/iosxr_router_ospf_vrf_redistribute_bgp/import.sh
@@ -1,1 +1,0 @@
-terraform import iosxr_router_ospf_vrf_redistribute_bgp.example "Cisco-IOS-XR-um-router-ospf-cfg:/router/ospf/processes/process[process-name=OSPF1]/vrfs/vrf[vrf-name=VRF1]/redistribute/bgp/as[as-number=65001]"