+++ examples/resources/iosxr_router_ospf_vrf_redistribute_isis/import.sh
@@ -1,1 +1,0 @@
-terraform import iosxr_router_ospf_vrf_redistribute_isis.example "Cisco-IOS-XR-um-router-ospf-cfg:/router/ospf/processes/process[process-name=OSPF1]/vrfs/vrf[vrf-name=VRF1]/redistribute/isis[instance-name=P1]"