+++ internal/provider/resource_iosxr_router_ospf_vrf_test.go
@@ -38,6 +38,22 @@ func TestAccIosxrRouterOSPFVRF(t *testing.T) {
 					resource.TestCheckResourceAttr("iosxr_router_ospf_vrf.test", "default_information_originate", "true"),
 					resource.TestCheckResourceAttr("iosxr_router_ospf_vrf.test", "default_information_originate_always", "true"),
 					resource.TestCheckResourceAttr("iosxr_router_ospf_vrf.test", "default_information_originate_metric_type", "1"),
+					resource.TestCheckResourceAttr("iosxr_router_ospf_vrf.test", "areas.0.area_id", "0"),
+					resource.TestCheckResourceAttr("iosxr_router_ospf_vrf.test", "redistribute_bgp.0.as_number", "65001"),
+					resource.TestCheckResourceAttr("iosxr_router_ospf_vrf.test", "redistribute_bgp.0.tag", "3"),
+					resource.TestCheckResourceAttr("iosxr_router_ospf_vrf.test", "redistribute_bgp.0.metric_type", "1"),
+					resource.TestCheckResourceAttr("iosxr_router_ospf_vrf.test", "redistribute_isis.0.instance_name", "P1"),
+					resource.TestCheckResourceAttr("iosxr_router_ospf_vrf.test", "redistribute_isis.0.level_1", "true"),
+					resource.TestCheckResourceAttr("iosxr_router_ospf_vrf.test", "redistribute_isis.0.level_2", "false"),
+					resource.TestCheckResourceAttr("iosxr_router_ospf_vrf.test", "redistribute_isis.0.level_1_2", "false"),
+					resource.TestCheckResourceAttr("iosxr_router_ospf_vrf.test", "redistribute_isis.0.tag", "3"),
+					resource.TestCheckResourceAttr("iosxr_router_ospf_vrf.test", "redistribute_isis.0.metric_type", "1"),
+					resource.TestCheckResourceAttr("iosxr_router_ospf_vrf.test", "redistribute_ospf.0.instance_name", "OSPF2"),
+					resource.TestCheckResourceAttr("iosxr_router_ospf_vrf.test", "redistribute_ospf.0.match_internal", "true"),
+					resource.TestCheckResourceAttr("iosxr_router_ospf_vrf.test", "redistribute_ospf.0.match_external", "false"),
+					resource.TestCheckResourceAttr("iosxr_router_ospf_vrf.test", "redistribute_ospf.0.match_nssa_external", "false"),
+					resource.TestCheckResourceAttr("iosxr_router_ospf_vrf.test", "redistribute_ospf.0.tag", "4"),
+					resource.TestCheckResourceAttr("iosxr_router_ospf_vrf.test", "redistribute_ospf.0.metric_type", "1"),
 				),
 			},
 			{
@@ -84,6 +100,30 @@ func testAccIosxrRouterOSPFVRFConfig_all() string {
 		default_information_originate = true
 		default_information_originate_always = true
 		default_information_originate_metric_type = 1
+		areas = [{
+			area_id = "0"
+		}]
+		redistribute_bgp = [{
+			as_number = "65001"
+			tag = 3
+			metric_type = "1"
+		}]
+		redistribute_isis = [{
+			instance_name = "P1"
+			level_1 = true
+			level_2 = false
+			level_1_2 = false
+			tag = 3
+			metric_type = "1"
+		}]
+		redistribute_ospf = [{
+			instance_name = "OSPF2"
+			match_internal = true
+			match_external = false
+			match_nssa_external = false
+			tag = 4
+			metric_type = "1"
+		}]
 	}
 	`
 }